[tune] Remove deprecated Resources class
