+++ python/ray/tune/tests/test_trial_runner.py
@@ -9,7 +9,6 @@ from ray.rllib import _register_all
 from ray import tune
 from ray.tune import TuneError, register_trainable
 from ray.tune.execution.ray_trial_executor import RayTrialExecutor
-from ray.tune.resources import Resources
 from ray.tune.schedulers import TrialScheduler, FIFOScheduler
 from ray.tune.search import BasicVariantGenerator
 from ray.tune.experiment import Trial
@@ -133,7 +132,7 @@ class TrialRunnerTest(unittest.TestCase):
             trial_executor=RayTrialExecutor(resource_manager=self._resourceManager())
         )
         kwargs = {
-            "resources": Resources(cpu=1, gpu=0.5),
+            "placement_group_factory": PlacementGroupFactory([{"CPU": 1, "GPU": 0.5}]),
         }
         trials = [
             Trial("__fake", **kwargs),
@@ -152,18 +151,11 @@ class TrialRunnerTest(unittest.TestCase):
         self.assertEqual(trials[2].status, Trial.PENDING)
         self.assertEqual(trials[3].status, Trial.PENDING)
 
-    def testResourceNumericalError(self):
-        resource = Resources(cpu=0.99, gpu=0.99, custom_resources={"a": 0.99})
-        small_resource = Resources(cpu=0.33, gpu=0.33, custom_resources={"a": 0.33})
-        for i in range(3):
-            resource = Resources.subtract(resource, small_resource)
-        self.assertTrue(resource.is_nonnegative())
-
     def testResourceScheduler(self):
         ray.init(num_cpus=4, num_gpus=1)
         kwargs = {
             "stopping_criterion": {"training_iteration": 1},
-            "resources": Resources(cpu=1, gpu=1),
+            "placement_group_factory": PlacementGroupFactory([{"CPU": 1, "GPU": 1}]),
         }
         trials = [Trial("__fake", **kwargs), Trial("__fake", **kwargs)]
 
@@ -185,7 +177,7 @@ class TrialRunnerTest(unittest.TestCase):
         ray.init(num_cpus=4, num_gpus=2)
         kwargs = {
             "stopping_criterion": {"training_iteration": 5},
-            "resources": Resources(cpu=1, gpu=1),
+            "placement_group_factory": PlacementGroupFactory([{"CPU": 1, "GPU": 1}]),
         }
         trials = [Trial("__fake", **kwargs), Trial("__fake", **kwargs)]
         snapshot = TrialStatusSnapshot()
@@ -209,7 +201,7 @@ class TrialRunnerTest(unittest.TestCase):
         )
         kwargs = {
             "stopping_criterion": {"training_iteration": 2},
-            "resources": Resources(cpu=1, gpu=0),
+            "placement_group_factory": PlacementGroupFactory([{"CPU": 1, "GPU": 0}]),
         }
         trials = [Trial("__fake", **kwargs)]
         for t in trials:
@@ -247,7 +239,7 @@ class TrialRunnerTest(unittest.TestCase):
         )
         kwargs = {
             "stopping_criterion": {"training_iteration": 2},
-            "resources": Resources(cpu=1, gpu=0),
+            "placement_group_factory": PlacementGroupFactory([{"CPU": 1, "GPU": 0}]),
         }
         trials = [Trial("__fake", **kwargs)]
         for t in trials:
